Crawl4ai/requirements.txt
# UPSC Content Crawler - Python Dependencies
# Core crawler
crawl4ai>=0.4.0
aiohttp>=3.9.0

# FastAPI Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.9

# Scheduling
schedule>=1.2.0

# Error handling and retries
backoff>=2.2.1

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# Utilities
beautifulsoup4>=4.12.0
lxml>=4.9.0
